requirements.txt
# AI Video Detective - 7B Model Only
# Core dependencies for GPU-powered local AI video analysis with streaming

# Core framework
flask>=2.3.0
python-dotenv>=1.0.0

# GPU acceleration
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# AI model framework
transformers>=4.40.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Qwen2.5-VL utilities
qwen-vl-utils[decord]==0.0.8

# Video processing
opencv-python-headless>=4.8.0
av>=10.0.0
decord>=0.6.0
ffmpeg-python>=0.2.0

# GPU optimization
flash-attn>=2.3.0
xformers>=0.0.22

# Streaming and real-time processing
websockets>=12.0.0
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0

# Utilities
numpy>=1.24.0
pillow>=10.0.0
psutil>=5.9.0

# Performance monitoring
pynvml>=11.5.0 